% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrb.R
\name{ctrb}
\alias{ctrb}
\title{Form Controllability Matrix}
\usage{
ctrb(A, B)
}
\arguments{
\item{A}{State matrix, A}

\item{B}{State matrix, B}
}
\value{
Returns the controllability matrix.
}
\description{
\code{ctrb} forms the controllability matrix.
}
\details{
\code{ctrb}  \code{ctrb(a, b)} returns the controllability matrix,  [B AB A^2B ... A^(n-1)B]. If the Controllability
matrix has full row rank, the system is controllable.
}
\examples{
a1 <- rbind(c(0,0),c(1,-3))
b1 <- rbind(-2,0)
ctrb(a1, b1)

}
\seealso{
\code{\link{obsv}}
}
