% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcgain.R
\name{dcgain}
\alias{dcgain}
\title{DC Gain}
\usage{
dcgain(sys)
}
\arguments{
\item{sys}{A transfer function or state-space model}
}
\value{
Returns the gain.
}
\description{
\code{dcgain} Forms the Givens rotation matrix
}
\details{
\code{dcgain(sys)}  Computes the steady-state gain (or low frequency gain) of a continuous
 system.
}
