% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlyap.R
\name{dlyap}
\alias{dlyap}
\title{Discrete Lyapunov Equation Solution}
\usage{
dlyap(a, c)
}
\arguments{
\item{a}{A square matrix}

\item{c}{A square matrix}
}
\value{
The functions returns a matrix
}
\description{
dlyap(A,C) solves the discrete Lyapunov equation:

A*X*A' + C = X
}
\details{
The solution X is symmetric when C is symmetric.
}
\examples{
D = rbind(c(1,2), c(2,1))
B = rbind(c(4,3), c(4,3))
dlyap(B,D)

}
\seealso{
\code{\link{lyap}}
}
