% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esort.R
\name{esort}
\alias{esort}
\title{Sort Complex Continuous Eigenvalues in Descending
Order}
\usage{
esort(p)
}
\arguments{
\item{p}{A vector containing the poles of a transfer-function, zero-pole model or the eigenvalues of a state-matrix}
}
\value{
Returns the sorted eigenvalues and the cooresponding
indices in a list:

   s   = sorted eigenvalues
   idx = index
}
\description{
\code{esort} sorts the complex continuous eigenvalues in descending
order
}
\details{
\code{esort} sorts the complex eigenvalues based on their real part. The unstable eigenvalues (positive
real part) are first shown.

This function is used to sort eigenvalues and system poles in \code{\link{damp}}
}
\seealso{
\code{\link{damp}}
}
