% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{givens_rot}
\alias{givens_rot}
\title{Complex Givens Rotation}
\usage{
givens_rot(a, b)
}
\arguments{
\item{a}{Complex Square-matrix}

\item{b}{complex Input-matrix}
}
\value{
Returns the complex Givens rotation matrix.
}
\description{
\code{givens_rot} Forms the Givens rotation matrix
}
\details{
\code{givens_rot(a, b)} returns the complex Givens rotation matrix
 This function is called by \code{\link{ordschur}}
}
\seealso{
\code{\link{ordschur}}
}
