% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issiso.R
\name{issiso}
\alias{issiso}
\alias{ismimo}
\title{SISO / MIMO Check}
\usage{
issiso(sys)
}
\arguments{
\item{sys}{Dynamic system of state-space model}
}
\value{
Returns TRUE or FALSE
}
\description{
\code{issiso} checks if state-space system is a single-input single-output system
\code{ismimo} checks if state-space system is a multiple-input multiple-output system
}
