% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyap.R
\name{lyap}
\alias{lyap}
\title{Continuous Lyapunov equation solution.}
\usage{
lyap(A, B, C = NULL)
}
\arguments{
\item{A}{An appropriately dimensioned matrix}

\item{B}{An appropriately dimensioned matrix}

\item{C}{An appropriately dimensioned matrix}
}
\value{
The functions returns a matrix
}
\description{
\code{lyap(A, B)} solves the particular Lyapunov equation:

A*X+X*A' = -B

 \code{lyap(A, B, C)} solves the Sylvester equation
}
\details{
This function solves the lyapunov equation using an eigenvalue
decomposition.
The solution X is symmetric when B is symmetric.
}
\note{
WARNING: This function assumes that there are no repeated eigenvalues.
      If repeated eigenvalues exist, then the solution may be
      unreliable.
}
\examples{
A = rbind(c(1,2), c(-3, -4))
Q = rbind(c(3,1), c(1,1))
lyap(A, Q)

}
\seealso{
\code{\link{dlyap}}
}
