% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsv.R
\name{obsv}
\alias{obsv}
\title{Observability Matrix}
\usage{
obsv(A, C)
}
\arguments{
\item{A}{State matrix, A}

\item{C}{State matrix, C}
}
\value{
\code{obsv(A, C)} returns the observability matrix.
where
     | C        |
     | CA       |
obsm =| CA^2     |
     | ...      |
     | CA^(n-1) |
}
\description{
This function creates the observability matrix.
}
\examples{
A <- rbind(c(0,1), c(-25,-4))
C <- rbind(c(1,0), c(0,1))
 obsv(A, C)

}
\seealso{
\code{\link{ctrb}}
}
