% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2str.R
\name{poly2str}
\alias{poly2str}
\title{Print Polynomial}
\usage{
poly2str(p, svar = "x", smul = "*", d = options("digits")$digits)
}
\arguments{
\item{p}{numeric vector representing a polynomial}

\item{svar}{character representing the unknown, default x.}

\item{smul}{multiplication symbol, default *.}

\item{d}{significant digits, default options("digits").}
}
\value{
Returns the usual string representing a polynomial in mathematics.
}
\description{
Print polynomial as a character string.
}
\details{
Modified from package *pracma*.
Modification: To hide any coefficient and power that is equal to 1
So that instead of '1s^3' we have 's^3' and instead of 's^1', we have 's'
}
\examples{
poly2str(c(2, -3, 1, 20, -11))

}
