% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFunction.R
\name{polysub}
\alias{polysub}
\title{Subtracting Polynomials}
\usage{
polysub(a, b)
}
\arguments{
\item{a}{Vector representing first polynomial.}

\item{b}{Vector representing second polynomial.}
}
\value{
Returns a Vector representing the resulting polynomial.
}
\description{
Subtract two polynomials given as vectors
}
\details{
Simply calls \code{polyadd} from \code{pracma} package
in the following manner: \code{pracma::polyadd(a, -b)}
}
\examples{
polysub(c(1, 1, 1), 1)
polysub(c(1, 1, 1), c(0, 0, 1))

}
