% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss2tf.R
\name{ss2tf}
\alias{ss2tf}
\title{State-space model conversion to Transfer function model.}
\usage{
ss2tf(a, b, c, d, iu) for systems with more than one input
ss2tf(a, b, c, d)
ss2tf(sys)
ss2tf(sys, iu) for systems with more than one input
}
\arguments{
\item{a}{An n x n matrix}

\item{b}{An n x m matrix}

\item{c}{An p x n matrix}

\item{d}{An p x m matrix}

\item{iu}{A numeric vector denoting number of inputs. default value is 1.}

\item{sys}{An object of state-space class}
}
\value{
Returns an object of 'tf' class containing \code{num} and \code{den}. The numerator coefficients
are returned in matrix num with as many rows as outputs y.
}
\description{
\code{ss2tf} converts the model for a state-space system to transfer function representation
}
\details{
\code{ss2tf} converts a model object in state-space form to transfer function model by calculating the transfer function of the system:
.
x = Ax + Bu
y = Cx + Du
}
\examples{
sys2 <- tf2ss(tf(1, c(1,2,1)))
ss2tf(sys2)
## OR
ss2tf(sys2$A,sys2$B,sys2$C,sys2$D)

}
\seealso{
\code{\link{tf2ss}} \code{\link{ss2zp}}
}
