% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss2zp.R
\name{ss2zp}
\alias{ss2zp}
\title{State-space representation to zero-pole-gain representation}
\usage{
ss2zp(a,b,c,d)
ss2zp(a,b,c,d,iu) #for SIMO systems
ss2zp(sys)
ss2zp(sys, iu) for SIMO systems
}
\arguments{
\item{a}{An n x n matrix}

\item{b}{An n x m matrix}

\item{c}{An p x n matrix}

\item{d}{An p x m matrix}

\item{sys}{An object of state-space class}
}
\value{
Returns a list object of 'zpk' class, consisting of z, p and k. The numerator zeros are returned in the columns of matrix Z with number of columns equal to number of outputs.  The gains for
 each numerator transfer function are returned in column vector K. P, a column vector contains the pole locations of the denominator of the transfer function.
}
\description{
\code{ss2zp} converts a system represented in state-space form to zero-pole-gain model
}
\details{
\code{ss2zp} converts a system represented in zero-pole form to state-space by converting from zero-pole to transfer function and from transfer functon to state-space
The vector P contains the pole locations of the denominator of the transfer function.
}
\examples{
A <- rbind(c(-2, -1), c(1,0)); B <- rbind(1,0);
C <- cbind(0,1); D <- 0;
sys2 <- ss(A,B,C,D)
ss2zp(sys2$A,sys2$B,sys2$C,sys2$D)
ss2zp( zp2ss ( tf2zp( c(1,1,1), c(1,2,1) ) ) )

}
\seealso{
\code{\link{zp2ss}} \code{\link{ss2tf}}
}
