% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_control_R6.R
\name{testClass}
\alias{testClass}
\title{Test Class for Control Statements}
\description{
A Test Class for Control Statements handling in R6
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-testClass-take_a_break}{\code{testClass$take_a_break()}}
\item \href{#method-testClass-next_up}{\code{testClass$next_up()}}
\item \href{#method-testClass-clone}{\code{testClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-testClass-take_a_break"></a>}}
\if{latex}{\out{\hypertarget{method-testClass-take_a_break}{}}}
\subsection{Method \code{take_a_break()}}{
Test if .parse_R6_expression () breaks with control function 'break'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{testClass$take_a_break()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-testClass-next_up"></a>}}
\if{latex}{\out{\hypertarget{method-testClass-next_up}{}}}
\subsection{Method \code{next_up()}}{
Test if .parse_R6_expression () breaks with control function 'next'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{testClass$next_up()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-testClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-testClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{testClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
