% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfchk.R
\name{tfchk}
\alias{tfchk}
\title{Transfer function check.}
\usage{
tfchk(num, den)
}
\arguments{
\item{num}{A numeric vector}

\item{den}{A numeric vector}
}
\value{
Returns a list with a proper transfer function or with warnings when not proper.
}
\description{
\code{tfchk} verifies the structure of a transfer function
}
\details{
This is a utility function that is always invoked by other functions to
verify the structure of \code{num, den}. Where \code{num} is the numerator and \code{den} is the denominator
of the transfer function. If the transfer function is not proper, it returns a list with length(num) = length(den).
}
\examples{
tf1 <- tfchk(1, c(1,2,1))

}
