% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{tf}
\alias{tf}
\title{Create Transfer function Model.}
\usage{
tf(num, den, Ts = NULL)
}
\arguments{
\item{num}{A numeric vector}

\item{den}{A numeric vector}
}
\value{
Returns an object of 'tf' class list with a proper transfer function or with warnings when not proper.
}
\description{
\code{tf} creates the model for a transfer function
}
\details{
\code{tf} creates a model object for transfer functions, Where \code{num} is the numerator and \code{den} is the denominator
of the transfer function. If the transfer function is not proper, it returns a list with length(num) = length(den).
}
\examples{
tf(1, c(1,2,1))
sys1 <- tf(1, c(1,2,1))
sys1$num
sys1$den

}
